refactor: create segmentation library and move segmentation code and tests
