+++ src/photon_mosaic/segmentation/__init__.py
@@ -1,0 +1,13 @@
+from .suite2p import (
+    Suite2pDetectedRois,
+    Suite2pEpochSegmentations,
+    Suite2pSegmentationSettings,
+    detect_rois_suite2p,
+)
+
+__all__ = [
+    "Suite2pDetectedRois",
+    "Suite2pEpochSegmentations",
+    "Suite2pSegmentationSettings",
+    "detect_rois_suite2p",
+]
